+++ examples/reader-ssr/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @olas/example-reader-ssr
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @kontsedal/olas-core@0.0.3
+  - @kontsedal/olas-react@0.0.3
+  - @kontsedal/olas-persist@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @kontsedal/olas-core@0.1.0
+  - @kontsedal/olas-react@1.0.0
+  - @kontsedal/olas-persist@1.0.0
+
 ## 0.0.1
 
 ### Patch Changes