+++ examples/stock-ticker/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @olas/example-stock-ticker
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @kontsedal/olas-core@0.0.3
+  - @kontsedal/olas-persist@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @kontsedal/olas-core@0.1.0
+  - @kontsedal/olas-persist@1.0.0
+
 ## 0.0.1
 
 ### Patch Changes